feat: finished main funcs of home screen
