+++ app/(tabs)/_layout.tsx
@@ -1,10 +1,7 @@
-import useUser from "@/hooks/useUser";
 import { Tabs } from "expo-router";
 import { Feather } from "@expo/vector-icons";
-import { Image, Text, View } from "react-native";
 
 export default function TabsLayout() {
-  const { isLoading, data } = useUser();
   return (
     <Tabs
       screenOptions={{
@@ -27,41 +24,6 @@ export default function TabsLayout() {
               <Feather
                 size={28}
                 name="home"
-                color={`${focused ? "white" : "#64748b"}`}
-              />
-            );
-          },
-          headerLeft: () => {
-            return (
-              <View className="pl-6">
-                <Image
-                  className="w-[75px] h-[75px]"
-                  source={require("@/assets/images/egoist-letter.png")}
-                />
-              </View>
-            );
-          },
-          headerRight: () => {
-            return (
-              <View className="pr-8">
-                <Text className="text-egoist-white text-xl font-bold">
-                  {isLoading ? "Loading..." : `${data?.goalWeight ?? 0} lbs`}
-                </Text>
-              </View>
-            );
-          },
-        }}
-      />
-      <Tabs.Screen
-        name="home2"
-        options={{
-          headerTitle: "",
-          headerTransparent: true,
-          tabBarIcon: ({ focused }) => {
-            return (
-              <Feather
-                size={28}
-                name="anchor"
                 color={`${focused ? "white" : "#64748b"}`}
               />
             );
